Fix recursive searching, fix COMM and TXXX metadata write errors
Search depth was always only 1 directory deeper than selected because the recursive property was not passed. This has now been fixed.

COMM (comment) could be set in weird and unsupported ways in the mp3 file. If we find a comment has already been set we clear it's data and then add the added with SongMetadataDownloader later.

TXXX (text) could sometimes be set in an unsupported way in an MP3 file we now remove this atribute as it is not needed or used usualy.